# src/main/resources/影院订票系统.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "影院订票系统",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/sessions/{sessionId}/seats": {
      "get": {
        "summary": "获取场次座位图",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "场次ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session_id": {
                      "type": "integer"
                    },
                    "seats": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "row_no": {
                            "type": "string"
                          },
                          "col_no": {
                            "type": "integer"
                          },
                          "type": {
                            "type": "string"
                          },
                          "price_factor": {
                            "type": "number"
                          },
                          "status": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "session_id",
                    "seats"
                  ]
                },
                "example": {
                  "session_id": "Long",
                  "seats": [
                    {
                      "id": "Long, 座位ID",
                      "row_no": "String, 行号",
                      "col_no": "Integer, 列号",
                      "type": "String, 座位类型",
                      "price_factor": "Double, 价格系数",
                      "status": "String, 状态(AVAILABLE/LOCKED)"
                    }
                  ]
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/movies/{movieId}": {
      "get": {
        "summary": "获取电影详情",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "description": "电影ID  ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "director": {
                      "type": "string"
                    },
                    "actors": {
                      "type": "string"
                    },
                    "duration": {
                      "type": "integer"
                    },
                    "release_date": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "poster_url": {
                      "type": "string"
                    },
                    "trailer_url": {
                      "type": "string"
                    },
                    "rating": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "director",
                    "actors",
                    "duration",
                    "release_date",
                    "description",
                    "poster_url",
                    "trailer_url",
                    "rating",
                    "status"
                  ]
                },
                "example": {
                  "id": "Long",
                  "title": "String",
                  "director": "String",
                  "actors": "String",
                  "duration": "Integer",
                  "release_date": "Date",
                  "description": "String",
                  "poster_url": "String",
                  "trailer_url": "String",
                  "rating": "Double",
                  "status": "String"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/movies": {
      "get": {
        "summary": "分页查询电影（后台管理）",
        "description": "支持按标题模糊搜索，分页返回电影列表。",
        "tags": [
          "电影",
          "后台"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码(默认1)",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页数量(默认10)",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "标题关键词",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回电影分页数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieSearchResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建电影信息",
        "description": "添加新电影（后台管理）",
        "tags": [
          "电影",
          "后台"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "电影创建成功"
          }
        }
      }
    },
    "/admin/movies/{movieId}": {
      "put": {
        "summary": "更新电影信息",
        "description": "修改电影基本信息或状态（后台管理）",
        "tags": [
          "电影",
          "后台"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        }
      },
      "delete": {
        "summary": "删除电影",
        "description": "永久删除电影（后台管理）",
        "tags": [
          "电影",
          "后台"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          }
        }
      }
    },
...
    "/admin/sessions": {
      "post": {
        "summary": "创建放映场次",
        "deprecated": false,
        "description": "为指定电影创建新场次",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "场次创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/sessions/{sessionId}": {
      "put": {
        "summary": "调整场次信息",
        "deprecated": false,
        "description": "修改场次时间、影厅或票价",
        "tags": [],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "取消场次",
        "deprecated": false,
        "description": "取消指定场次并通知相关用户",
        "tags": [],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "取消成功",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/sessions/{sessionId}/seats": {
      "get": {
        "summary": "监控场次售票情况",
        "deprecated": false,
        "description": "获取场次座位销售状态",
        "tags": [],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "座位状态图",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatMapResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/movies/today": {
      "post": {
        "summary": "设置今日上映电影",
        "description": "管理员设置今日上映电影，body为电影ID数组。",
        "tags": [
          "电影",
          "后台"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer"
                }
              },
              "example": [
                1,
                5,
                8,
                12
              ]
            }
          }
        },
        "responses": {
          "200": {
            "description": "设置成功"
          }
        }
      },
      "get": {
        "summary": "获取今日上映电影ID列表",
        "description": "管理员获取当前今日上映电影ID数组。",
        "tags": [
          "电影",
          "后台"
        ],
        "responses": {
          "200": {
            "description": "返回今日上映电影ID数组",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                "example": [
                  1,
                  5,
                  8,
                  12
                ]
              }
            }
          }
        }
      }
    },
    "/api/movies/today": {
      "get": {
        "summary": "获取今日上映电影详情",
        "description": "分页获取管理员设置的今日上映电影详情列表。",
        "tags": [
          "电影"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码(默认1)",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页数量(默认10)",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回今日上映电影分页数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieSearchResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/admin/sessions/auto-arrange": {
      "post": {
        "summary": "自动排片",
        "description": "为今日上映电影自动安排场次，评分高优先，黄金时段优先，避免同厅时间冲突。body为电影ID数组。返回安排成功的场次数。",
        "tags": [
          "场次",
          "后台"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer"
                }
              },
              "example": [
                1,
                5,
                8,
                12
              ]
            }
          }
        },
        "responses": {
          "200": {
            "description": "安排成功的场次数",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                },
                "example": 5
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "电影ID"
          },
          "title": {
            "type": "string",
            "description": "电影标题"
          },
          "director": {
            "type": "string",
            "description": "导演"
          },
          "actors": {
            "type": "string",
            "description": "演员列表"
          },
          "duration": {
            "type": "integer",
            "description": "时长(分钟)"
          },
          "release_date": {
            "type": "string",
            "format": "date",
            "description": "上映日期"
          },
          "description": {
            "type": "string",
            "description": "剧情简介"
          },
          "poster_url": {
            "type": "string",
            "description": "海报URL"
          },
          "trailer_url": {
            "type": "string",
            "description": "预告片URL"
          },
          "rating": {
            "type": "number",
            "format": "float",
            "description": "评分"
          },
          "status": {
            "type": "string",
            "enum": [
              "UPCOMING",
              "SHOWING",
              "OFF"
            ],
            "description": "上映状态"
          }
        }
      },
...
      "CreateSessionRequest": {
        "type": "object",
        "required": [
          "movieId",
          "hallId",
          "startTime",
          "price"
        ],
        "properties": {
          "movieId": {
            "type": "integer",
            "description": "电影ID"
          },
          "hallId": {
            "type": "integer",
            "description": "影厅ID"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "description": "开场时间"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "基础票价"
          }
        }
      },
...
      "UpdateSessionRequest": {
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string",
            "format": "date-time",
            "description": "新开场时间"
          },
          "hallId": {
            "type": "integer",
            "description": "新影厅ID"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "新票价"
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "场次ID"
          },
          "movieId": {
            "type": "integer",
            "description": "电影ID"
          },
          "hallId": {
            "type": "integer",
            "description": "影厅ID"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "description": "开场时间"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "description": "结束时间"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "基础价格"
          },
          "available_seats": {
            "type": "integer",
            "description": "可用座位数"
          }
        }
      },
...
      "SeatMapResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer"
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "rowNo": {
                  "type": "string"
                },
                "seats": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeatStatus"
                  }
                }
              }
            }
          }
        }
      },
      "SeatStatus": {
        "type": "object",
        "properties": {
          "colNo": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "LOCKED",
              "OCCUPIED"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}